Occurrences responsible for updating accumulators.
When a Occurrence object is passed to a Portfolio object
the method Occurrence.update_portfolio is called. This
method receives the Portfolio object itself, and its up the
Occurrence object to perform any needed operation.

When the Occurrence object is passed to the Accumulator,
the method update_accumulator is called. This method receives
the Accumulator object itself, and its up the Operation
object to update the accumulator.

Accumulators now have no need to know anything about the
objects they are accumulating. Accumulators now save their
current state in a attribute named data, a dictionary that
should be handled by whatever asset or operation the Accumulator
is dealing with.

Operation, OperationContainer and all functionalities related
to Operation are on the operations.py module.