Asset as a subclass of Subject.
Subject is defined in trade.py and is used to represent a subject
of an Occurrence.

Subjects are responsible for setting the initial state of the
Accumulator that is accumulating them.

Added the default expiration functionality do Subjects; it resets
the data on the Accumulator to the initial state of the Subject.